actor.lisp: Extracted the majority of the actor class into an event-mixin. Ignore-this: 9565f9c71f1dbb122176a8e2c90073d7
The shared components of the actor and event-subscription classes can be implemented in the event-mixin class.

Instead of having specific handle-key-down and handle-key-up events in the game, game inherits from event-subscription and gets push-event calls directly. The key-subscription object then subscribes to the game and sub-delegates just key events. Although it would be entirely possible to remove the key-subscription object, keeping it as an intermediate layer has the advantage that it can be used to implement key-modes (e.g. play vs repl).

The actor class is now a trivial class which calls dispatch-queued-events in its event-update method.