load.lisp, main.lisp: Debug mode allows code to be edited interactively. Ignore-this: a4bbe3617d9b00b58b1532ecba501805
When :blt-debug is in *features*, main will automatically check for changed files approximately every 0.5 seconds. Also, load.lisp automatically enables debug mode so that users can continue editing their code as the game is running.

Of course, proper use of this feature requires that user code keep all state in the game and component tree, and adapt well to any changes they are likely to make.