load.lisp, main.lisp: Removed debug mode. Ignore-this: 4150e3347262711c62d77c8b55f16b77
The automatic reloading of files was causing intermittent pauses even when no changes had to be loaded. So in the interest of maintaining a consistent development experience, the automatic reload check has been removed. And since that check was the only reason why debug mode existed at all, debug mode was also removed.

Reloading files may still be desirable, however. One option for this might be to set a key which checks for changed files. Another (probably more difficult) approach would be to register a handler with the OS to get events when the files are changed (so the game loop doesn't have to check explicitly).