Improved interactive development: Re-added reload feature.
When starting the game via load.lisp, the engine automagically installs a key binding for the R key which reloads the game. While this is somewhat magical, it is a useful feature for development. The feature is only compiled into the game when blt-debug is in *features*.

Also fixed a resource directory issue for Mac deployment.