Update is now an event.
Non-actor objects are now considered immutable. The only way to update an actor object is through its events. Now that update is an event, all traces of the update/event-update split have been removed. (Actually, event-update is still around, but it exists only to enqueue an update event for each actor.)